Move feeCalculations alongside classes using it
